backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-hcl2==4.3.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
xgboost==2.0.3
joblib==1.3.2
pydantic==2.5.2
